% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19SeverityPrediction3.R
\docType{package}
\name{Covid19SeverityPrediction3}
\alias{Covid19SeverityPrediction3}
\alias{Covid19SeverityPrediction3-package}
\title{Covid19SeverityPrediction3}
\description{
Covid19SeverityPrediction3
}
